+++ clash-protocols/src/Protocols/Experimental/Avalon/MemMap.hs
@@ -99,7 +99,8 @@ import Clash.Prelude qualified as C
 
 -- me
 
-import Protocols.DfConv qualified as DfConv
+import Protocols.Experimental.DfConv qualified as DfConv
+import Protocols.Experimental.Simulate
 import Protocols.Idle
 import Protocols.Internal
