+++ clash-protocols/src/Protocols/Experimental/Axi4/WriteData.hs
@@ -40,6 +40,7 @@ import Clash.Prelude qualified as C
 
 -- me
 import Protocols.Experimental.Axi4.Common
+import Protocols.Experimental.Simulate
 import Protocols.Idle
 import Protocols.Internal
