+++ clash-protocols/src/Protocols/Experimental/Axi4/WriteResponse.hs
@@ -27,6 +27,7 @@ import Clash.Prelude qualified as C
 
 -- me
 import Protocols.Experimental.Axi4.Common
+import Protocols.Experimental.Simulate
 import Protocols.Idle
 import Protocols.Internal
