+++ clash-protocols/tests/Tests/Protocols/PacketStream/Converters.hs
@@ -13,8 +13,8 @@ import Test.Tasty.Hedgehog (HedgehogTestLimit (HedgehogTestLimit))
 import Test.Tasty.Hedgehog.Extra (testProperty)
 import Test.Tasty.TH (testGroupGenerator)
 
-import Protocols.Hedgehog
-import Protocols.PacketStream.Converters
+import Protocols.Experimental.Hedgehog
+import Protocols.Experimental.PacketStream
 import Protocols.PacketStream.Hedgehog
 
 generateUpConverterProperty ::
