+++ clash-protocols/tests/Tests/Protocols/PacketStream/Packetizers.hs
@@ -18,9 +18,8 @@ import Test.Tasty.TH (testGroupGenerator)
 
 import Protocols
 import Protocols.Df qualified as Df
-import Protocols.Hedgehog
-import Protocols.PacketStream (packetizeFromDfC, packetizerC)
-import Protocols.PacketStream.Base
+import Protocols.Experimental.Hedgehog
+import Protocols.Experimental.PacketStream
 import Protocols.PacketStream.Hedgehog
 
 {- |
